requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crispy-forms==1.12.0
django-import-export==2.6.0
django-timezone-field==4.2.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
httpie==2.5.0
idna==3.2
importlib-metadata==4.8.3
kombu==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl
packaging==21.3
Pillow==8.3.2
prompt-toolkit==3.0.30
psycopg2==2.9.1
Pygments==2.10.0
pyparsing==3.0.7
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
redis==4.3.4
reportlab==3.6.5
requests==2.26.0
requests-toolbelt==0.9.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
